refactor: custom scribble notifier ->  mixin, extension 으로 기능 분리
