+++ lib/features/canvas/mixins/tool_management_mixin.dart
@@ -1,0 +1,41 @@
+import 'package:scribble/scribble.dart';
+
+import '../models/tool_mode.dart';
+
+/// 도구 관리 기능을 제공하는 Mixin
+mixin ToolManagementMixin on ScribbleNotifier {
+  ToolMode get toolMode;
+  set toolMode(ToolMode value);
+
+  /// 공통 도구 변경 메서드
+  void setTool(ToolMode newToolMode) {
+    toolMode = newToolMode;
+
+    if (newToolMode.isDrawingMode) {
+      temporaryValue = ScribbleState.drawing(
+        sketch: value.sketch,
+        selectedColor: newToolMode.defaultColor.toARGB32(),
+        selectedWidth: newToolMode.defaultWidth,
+        allowedPointersMode: value.allowedPointersMode,
+        scaleFactor: value.scaleFactor,
+        activePointerIds: value.activePointerIds,
+      );
+    } else {
+      // 지우개 모드
+      temporaryValue = ScribbleState.erasing(
+        sketch: value.sketch,
+        selectedWidth: newToolMode.defaultWidth,
+        scaleFactor: value.scaleFactor,
+        allowedPointersMode: value.allowedPointersMode,
+        activePointerIds: value.activePointerIds,
+      );
+    }
+  }
+
+  void setPen() => setTool(ToolMode.pen);
+  void setHighlighter() => setTool(ToolMode.highlighter);
+  void setLinker() => setTool(ToolMode.linker);
+
+  @override
+  void setEraser() => setTool(ToolMode.eraser);
+}
