--- lib/features/canvas/models/custom_scribble_notifier.dart
@@ -1,70 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:scribble/scribble.dart';
-
-import '../../../models/page.dart' as page_model;
-import 'tool_mode.dart';
-
-class CustomScribbleNotifier extends ScribbleNotifier {
-  CustomScribbleNotifier({
-    super.sketch,
-    super.allowedPointersMode,
-    super.maxHistoryLength,
-    super.widths = const [1, 3, 5, 7],
-    super.pressureCurve,
-    super.simplifier,
-    super.simplificationTolerance,
-    required this.canvasIndex,
-    required this.toolMode,
-    this.page, // 멀티페이지용 Page 객체 (선택사항)
-  });
-
-  final int canvasIndex;
-  ToolMode toolMode;
-  final page_model.Page? page; // 멀티페이지에서 사용할 Page 객체
-
-  /// 자동저장 구현
-  @override
-  void onPointerUp(PointerUpEvent event) {
-    super.onPointerUp(event);
-    saveSketch();
-  }
-
-  void saveSketch() {
-    // 멀티페이지 - Page 객체가 있으면 해당 Page에 저장
-    if (page != null) {
-      page!.updateFromSketch(currentSketch);
-    }
-  }
-
-  /// 공통 도구 변경 메서드
-  void _setTool(ToolMode newToolMode) {
-    toolMode = newToolMode;
-
-    if (newToolMode.isDrawingMode) {
-      temporaryValue = ScribbleState.drawing(
-        sketch: value.sketch,
-        selectedColor: newToolMode.defaultColor.toARGB32(),
-        selectedWidth: newToolMode.defaultWidth,
-        allowedPointersMode: value.allowedPointersMode,
-        scaleFactor: value.scaleFactor,
-        activePointerIds: value.activePointerIds,
-      );
-    } else {
-      // 지우개 모드
-      temporaryValue = ScribbleState.erasing(
-        sketch: value.sketch,
-        selectedWidth: newToolMode.defaultWidth,
-        scaleFactor: value.scaleFactor,
-        allowedPointersMode: value.allowedPointersMode,
-        activePointerIds: value.activePointerIds,
-      );
-    }
-  }
-
-  void setPen() => _setTool(ToolMode.pen);
-  void setHighlighter() => _setTool(ToolMode.highlighter);
-  void setLinker() => _setTool(ToolMode.linker);
-
-  @override
-  void setEraser() => _setTool(ToolMode.eraser);
-}
