+++ lib/features/canvas/widgets/drawing_mode_toolbar.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:scribble/scribble.dart';
 
-import '../models/custom_scribble_notifier.dart';
 import '../models/tool_mode.dart';
+import '../notifiers/custom_scribble_notifier.dart';
 
 /// 그리기 모드 툴바
 ///
